freeproxy/requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
lxml==4.3.3
PyExecJS==1.5.1
pymongo==3.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
